Additional tweaks to AI combat behaviour. Added the 'ScriptVar' implementation for storing/ reading script related variables.
Fundementally reworked some base classes and added the ScriptVar class
for reading variables in a way not too disimilar from the cvar_t
implementation of the quake engine.

+++ ScriptBase/Helpers/LoopedParticle.cs
@@ -31,7 +31,7 @@ namespace AirSuperiority.ScriptBase.Helpers
         /// <summary>
         /// Set the particle FX colour.
         /// </summary>
-        public Color Colour { set { Function.Call(Hash.SET_PARTICLE_FX_LOOPED_COLOUR, Handle, value.R, value.G, value.B, 0); } }
+        public Color Colour { set { Function.Call(Hash.SET_PARTICLE_FX_LOOPED_COLOUR, Handle, (float) value.R, (float) value.G, (float) value.B, 0); } }
 
         public LoopedParticle(string assetName, string fxName)
         {
